parse: const_expr_val results must be int64_t at all call sites
Sibling family of the case-statement fix in 9e6b44c. const_expr_val
returns int64_t, but four other call sites stored the result in `int`,
sign-extending values with bit 31 set when assigned to a wider field
or used in 64-bit arithmetic.

Sites:
  - aligned attribute   (parse.c:1451) — `int a` → `int64_t a`
  - vector_size attr    (parse.c:1545) — `int vs` → `int64_t vs`
  - array range init    (parse.c:4978) — `int lo,hi,j` → `int64_t`
  - init-with-deduce    (parse.c:5211) — `int lo,hi` → `int64_t`

None of these are as visibly broken as the switch-case site (no real
NetBSD code uses 0x80000000+ alignments or vector sizes), but the
class of bug is identical and the fix is a one-token change per site.
Worth landing as a hygiene pass so the family is closed.

Regression test in 19_const_expr_truncation_in_attrs.c. 19/19 regression,
30/30 compliance, stage1 == stage2 fixed point holds.

+++ src/parse.c
@@ -1448,7 +1448,7 @@ static Token *attribute_list(Token *tok, Type *ty, VarAttr *attr) {
       } else if (equal(tok, "aligned") || equal(tok, "__aligned__")) {
         tok = tok->next;
         if (consume(&tok, tok, "(")) {
-          int a = const_expr_val(&tok, tok);
+          int64_t a = const_expr_val(&tok, tok);
           tok = skip(tok, ")");
           if (ty) {
             ty->align = a;
@@ -1542,7 +1542,7 @@ static Token *attribute_list(Token *tok, Type *ty, VarAttr *attr) {
       } else if (equal(tok, "vector_size") || equal(tok, "__vector_size__")) {
         tok = tok->next;
         tok = skip(tok, "(");
-        int vs = const_expr_val(&tok, tok);
+        int64_t vs = const_expr_val(&tok, tok);
         tok = skip(tok, ")");
         if (attr) attr->vector_size = vs;
         // If ty is provided directly and we're NOT in declspec (attr is NULL),
@@ -4975,17 +4975,17 @@ static void array_initializer1(Token **rest, Token *tok, Initializer *init) {
 
     // Designated initializer (may jump backward)
     if (equal(tok, "[")) {
-      int lo = const_expr_val(&tok, tok->next);
+      int64_t lo = const_expr_val(&tok, tok->next);
       // Range designator: [lo ... hi] = val
       if (equal(tok, "...")) {
-        int hi = const_expr_val(&tok, tok->next);
+        int64_t hi = const_expr_val(&tok, tok->next);
         tok = skip(tok, "]");
         if (!consume(&tok, tok, "="))
           ; // = is optional in some modes
         // Parse the value via initializer2 for each element (handles
         // both scalar "= 42" and struct "= { .f = v }" forms).
         Token *val_start = tok;
-        for (int j = lo; j <= hi; j++) {
+        for (int64_t j = lo; j <= hi; j++) {
           tok = val_start;
           if (j < init->ty->array_len)
             initializer2(&tok, tok, init->children[j]);
@@ -5208,9 +5208,9 @@ static int count_array_init_elements(Token *tok, Type *ty) {
       if (is_end(tok)) break;
     }
     if (equal(tok, "[")) {
-      int lo = const_expr_val(&tok, tok->next);
+      int64_t lo = const_expr_val(&tok, tok->next);
       if (equal(tok, "...")) {
-        int hi = const_expr_val(&tok, tok->next);
+        int64_t hi = const_expr_val(&tok, tok->next);
         tok = skip(tok, "]");
         consume(&tok, tok, "=");
         initializer2(&tok, tok, dummy);
